[chore] Restructure files by feature
