+++ localtest/reset_faces.py
